+++ CHANGELOG.md
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Function `valid_server_proof?/6`.
 - Support all prime group from [RFC 5054].
 
-[Unreleased]: https://github.com/thiamsantos/srp-elixir/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/thiamsantos/srp-elixir/compare/v0.1.1...HEAD
 [0.1.0]: https://github.com/thiamsantos/srp-elixir/tree/v0.1.0
 [0.1.1]: https://github.com/thiamsantos/srp-elixir/tree/v0.1.1
 [stream_data]: https://hex.pm/packages/stream_data