+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.1.1] - 2018-11-03
-## Added
+### Added
 - Mix metadata.
 
 ## [0.1.0] - 2018-11-03